+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsellm.py
@@ -252,9 +252,9 @@ class EvaluatorResponseLlm(BaseModel):
 
     mode: Mode
 
-    created: Optional[str] = "2026-08-25T13:28:34.194Z"
+    created: Optional[str] = "2026-08-25T14:44:12.908Z"
 
-    updated: Optional[str] = "2026-08-25T13:28:34.194Z"
+    updated: Optional[str] = "2026-08-25T14:44:12.908Z"
 
     updated_by_id: OptionalNullable[str] = UNSET