+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-08-25T13:28:32.932Z")
+    updated: Optional[datetime] = parse_datetime("2026-08-25T14:44:11.279Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")