+++ packages/orq-rc/RELEASES.md
@@ -10479,3 +10479,13 @@ Based on:
 - [python v4.14.0-rc.54] packages/orq-rc
 ### Releases
 - [PyPI v4.14.0-rc.54] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.54 - packages/orq-rc
+
+## 2026-08-25 14:47:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.3 (2.932.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.14.0-rc.55] packages/orq-rc
+### Releases
+- [PyPI v4.14.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.14.0-rc.55 - packages/orq-rc